lambda_functions/textraction_lambda/requirements.txt
aws_lambda_powertools
amazon-textract-textractor
pdfplumber
pydantic
boto3
